jpa working

+++ src/main/java/com/example/common/NutritionCrud.java
@@ -7,7 +7,7 @@ import com.example.domain.Nutrition;
 public interface NutritionCrud {
 	Nutrition add(Nutrition nutrition);
 	List<Nutrition> findAll();
-	Nutrition find(int id);
+	Nutrition find(long id);
 	void update(Nutrition nutrition);
 	void delete(long id);
 	void delete(List<Long> ids);
