+++ src/main/java/com/example/controller/NutritionController.java
@@ -52,7 +52,7 @@ public class NutritionController {
 	StorageService storageService;
 
 	@GetMapping("/view-nutrition/{id}/files")
-	public String uploadFile(@PathVariable("id") Integer id, Model model) {
+	public String uploadFile(@PathVariable("id") Long id, Model model) {
 //		if (id == 0) {
 //			throw new RuntimeException("I'll be back");
 //		}
@@ -60,7 +60,7 @@ public class NutritionController {
                 .loadAll(id)
                 .map(path ->
                         MvcUriComponentsBuilder
-                                .fromMethodName(NutritionController.class, "serveFile", Integer.toString(id), path.getFileName().toString())
+                                .fromMethodName(NutritionController.class, "serveFile", Long.toString(id), path.getFileName().toString())
                                 .build().toString())
                 .collect(Collectors.toList()));
 		model.addAttribute("id", id);
@@ -69,7 +69,7 @@ public class NutritionController {
 	
 	 @GetMapping("/view-nutrition/{id}/files/{filename:.+}")
 	    @ResponseBody
-	    public ResponseEntity<Resource> serveFile(@PathVariable("id") Integer id, @PathVariable String filename) {
+	    public ResponseEntity<Resource> serveFile(@PathVariable("id") Long id, @PathVariable String filename) {
 			//model.addAttribute("id", id);
 	        Resource file = storageService.loadAsResource(id, filename);
 	        return ResponseEntity
@@ -81,7 +81,7 @@ public class NutritionController {
 	
 	 
 	@PostMapping("/view-nutrition/{id}/files")
-    public String handleFileUpload(@RequestParam("file") MultipartFile file,@PathVariable("id") Integer id,Model model,
+    public String handleFileUpload(@RequestParam("file") MultipartFile file,@PathVariable("id") Long id,Model model,
                                    RedirectAttributes redirectAttributes) {
         storageService.store(id, file);
         	nutritionService.addFile(id, file.getOriginalFilename());
