+++ src/main/java/com/example/dao/FileDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.domain.File;
 
-public interface FileDao extends JpaRepository<File, Integer>{
+public interface FileDao extends JpaRepository<File, Long>{
 
 }
