+++ src/main/java/com/example/dao/NutritionDao.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.domain.Nutrition;
 
 
-public interface NutritionDao extends JpaRepository<Nutrition, Integer> {
+public interface NutritionDao extends JpaRepository<Nutrition, Long> {
 }
