+++ src/main/java/com/example/service/NutritionService.java
@@ -4,5 +4,5 @@ import com.example.common.NutritionCrud;
 import com.example.domain.Nutrition;
 
 public interface NutritionService extends NutritionCrud{
-	Nutrition addFile(Integer id, String fileName);
+	Nutrition addFile(Long id, String fileName);
 }
